feat: full food CRUD, Vue refactor with Naive UI, package updates
API (.NET):
- Fix Food.Calorie and FoodInfo.Calorie type from int to decimal
- Add UpdateFullFoodAsync (replace nutrients collection atomically)
- Add DELETE endpoint for foods
- Add IsBrandInUse endpoint and service/repository method

Vue (WebVue):
- Update all packages to latest (Vue 3.5.30, Naive UI 2.44.1, Vite 8.0.0)
- FoodDetail: rewrite with n-form/n-form-item/n-grid, add missing Calorie field, n-spin loader
- FoodNutrientInput: rewrite with n-grid, visible nutrient name label
- FoodsList: add delete button and emit, clean up component
- BrandsList: remove UUID column and hardcoded scroll-x
- FoodsView: add handleDeleteFood with confirmation dialog
- App.vue: fix navigation with active state highlight, remove nested RouterLink pattern
- useAsyncState: discriminate error messages by HTTP status code (401/403/404/422/500)
- foods.api.ts: add deleteFood
- useFoods.ts: add removeFood

+++ src/Dr.NutrizioNino.Api/Infrastructure/Models/Food.cs
@@ -17,7 +17,7 @@ public partial class Food
 
     public Guid? BrandId { get; set; }
 
-    public int Calorie { get; set; }
+    public decimal Calorie { get; set; }
     public Guid UnitOfMeasureId { get; set; }
 
     public virtual Brand Brand { get; set; }
